Pass validation arguments to all field-editors
