+++ packages/core-types/addon/styles/addon.css
@@ -1,10 +1,8 @@
-.integer-editor.invalid > input {
+.field-editor.invalid > input {
   background-color: #FBC0BB;
 }
-.string-editor.invalid > input {
-  background-color: #FBC0BB;
-}
-.string-editor > .error-message {
+.field-editor > .error-message {
   color: red;
   font-size: 0.8rem;
 }
+
