+++ packages/core-types/addon/templates/components/field-editors/integer-editor.hbs
@@ -1,4 +1,4 @@
-<div class="integer-editor {{if invalid 'invalid'}}">
+<div class="field-editor {{if invalid 'invalid'}}">
 {{cs-one-way-input
        value=(get content field)
        update=(action "update")
